dev: create `AdaptyPlugin` contract to invert dependencies
